Allow updating multiple ECS services from a manifest
We include a fair bit of CLI-parsing code to make it easy to work with
both standalone `docker-compose.yml` files and manifests with multiple
services, and to preserve `docker-compose`-like command line conventions
as much as possible.

Basically, we want this to be a good tool, one which we use heavily for a
variety of tasks, and one which offers an easy migration path for people
familiar with `docker-compose`.