Add support for docker-compose-like `--project-name`
Right now, the only way to specify a prefix for all service names is to
generate it as part of a the manifest.  But this means that it's
impossible to run multiple copies of the same manifest under different
names.

This patch allows us to set a prefix for all task definition and service
names, turning "foo" into "staging-foo".