Fix JSON normalization
We keep encountering edge cases in the JSON normalization thanks to the
difficulty of sorting mixed data types with Ruby.  Rather than continue
to mess with this, I'm proposing a comprehensive fix with extensive
tests.  We define a new `compare_any` function that works like `<=>` but
which is defined to compare any two JSON values.

We test this providing many examples of how different pairs of values
should be ordered, and verifying the `compare_any` handles each the way
we want.  Then, using this improved comparison primitive, we rebuild
`sort_recursively`.

I've merged in the extra `sort_recursively` tests provided by Derek and
Seamus.

With luck, this should handle all the different JSON we need to
normalize for comparison.  And if there are still problematic cases,
well, we have a much stronger base to build on.