Make vault a regular dependency
We could try making this a regular dependency instead of a dev dependency. I think there's some magic that tries to do the right thing if vault isn't installed, so that this didn't need to always be enabled (or something like that), but I don't have the time to figure it out. Anyway, it's possible this might help.

CC @seamusabshere @dkastner